Validation: creation of remoteprefix if it is not yet existent
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@3734 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ validation/validationRun/validationRun.py
@@ -187,6 +187,16 @@ def doRun(directory, MardynExe):
         rsyncremote = remote
         if remote.endswith('-mic0') or remote.endswith('-mic1'):
             rsyncremote = remote[:-5]
+        command = "mkdir -p " + remoteprefix    
+        mkdircmd = []
+        mkdircmd.extend(['ssh', rsyncremote, command])
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE)
+        out, err = p.communicate()
+        if p.returncode:
+            print "error on mkdir -p:"
+            print out, err
+            exit(-1)        
+        
         remotedirectory = remoteprefix + "/" + directory
         command = "rsync --delete-before -r ../" + directory + " " + rsyncremote + ":" + remoteprefix
         print command
